+++ be/src/io/cache/block_file_cache_factory.h
@@ -25,8 +25,8 @@
 #include <vector>
 
 #include "common/status.h"
-#include "io/cache/block/block_file_cache.h"
-#include "io/cache/block/block_file_cache_settings.h"
+#include "io/cache/block_file_cache.h"
+#include "io/cache/block_file_cache_settings.h"
 namespace doris {
 class TUniqueId;
