+++ be/src/io/fs/buffered_reader.h
@@ -28,7 +28,7 @@
 #include <vector>
 
 #include "common/status.h"
-#include "io/cache/block/cached_remote_file_reader.h"
+#include "io/cache/cached_remote_file_reader.h"
 #include "io/file_factory.h"
 #include "io/fs/broker_file_reader.h"
 #include "io/fs/file_reader.h"
