+++ be/src/io/fs/s3_file_bufferpool.cpp
@@ -25,8 +25,8 @@
 #include "common/logging.h"
 #include "common/status.h"
 #include "common/sync_point.h"
-#include "io/cache/block/block_file_cache_fwd.h"
-#include "io/cache/block/block_file_segment.h"
+#include "io/cache/block_file_cache_fwd.h"
+#include "io/cache/block_file_segment.h"
 #include "io/fs/s3_common.h"
 #include "runtime/exec_env.h"
 #include "runtime/thread_context.h"
