+++ be/test/io/cache/file_block_cache_test.cpp
@@ -37,10 +37,10 @@
 
 #include "common/config.h"
 #include "gtest/gtest_pred_impl.h"
-#include "io/cache/block/block_file_cache.h"
-#include "io/cache/block/block_file_cache_settings.h"
-#include "io/cache/block/block_file_segment.h"
-#include "io/cache/block/block_lru_file_cache.h"
+#include "io/cache/block_file_cache.h"
+#include "io/cache/block_file_cache_settings.h"
+#include "io/cache/block_file_segment.h"
+#include "io/cache/block_lru_file_cache.h"
 #include "io/fs/path.h"
 #include "olap/options.h"
 #include "util/slice.h"
